Remove usage of the unstable, to-be-deprecated mpsc_select feature
The `select!` macro used in the tests in this repository is unstable
(only available on Nightly), but doesn’t have a clear path to stabilization
because of significant issues in its design:
https://github.com/rust-lang/rust/issues/27800.

The Rust Library Team is considering this API for deprecation and eventual
removal. We found it used in this repository through Crater in
https://github.com/rust-lang/rust/pull/54228.

In this case, it turns out that selection over two channels can be replaced
with waiting on a single channel with a custom `enum` for the message type,
and cloning the sender for sending from two different places/threads.
This PR implements that.

If you truely need selection over multiple channels, consider using
the [crossbeam-channel](https://crates.io/crates/crossbeam-channel) library.

If you feel that this feature should not be removed from the standard library,
please comment on the tracking issue:
https://github.com/rust-lang/rust/issues/27800